+++ Integrian3D/Components/CameraComponent/CameraComponent.h
@@ -9,12 +9,13 @@ namespace Integrian3D
 	class CameraComponent final
 	{
 	public:
-		CameraComponent(const float nearPlane, const float farPlane, const float fov, const float aspectRatio, const float speed = 10.f);
+		CameraComponent(const float nearPlane, const float farPlane, const float fov,
+			const float aspectRatio, const double speed = 10.0);
 
 		void SetView(const glm::mat4& view);
 
 		void SetSpeed(const float speed);
-		__NODISCARD float GetSpeed() const { return Speed; }
+		__NODISCARD double GetSpeed() const { return Speed; }
 
 		__NODISCARD const glm::mat4& GetView() const { return View; }
 		__NODISCARD const glm::mat4& GetProjection() const { return Projection; }
@@ -24,7 +25,7 @@ namespace Integrian3D
 		float FarPlane;
 		float FOV;
 		float AspectRatio;
-		float Speed;
+		double Speed;
 
 		glm::mat4 View;
 		glm::mat4 Projection;
